requirements.txt
nose
mock
requests
redis
pyserial
https://github.com/gareth-lloyd/pegpy/tarball/master
